New bluedot fixes.

+++ src/sagan-config.h
@@ -231,13 +231,14 @@ struct _SaganConfig
     char         bluedot_auth[64];
     char         bluedot_cat[MAXPATH];
     int          bluedot_timeout;
-    uint64_t     bluedot_ip_max_cache;
+    int	         bluedot_ip_max_cache;
     uint64_t	 bluedot_hash_max_cache;
     uint64_t	 bluedot_url_max_cache;
     uint64_t 	 bluedot_filename_max_cache;
     uint64_t     bluedot_last_time;                    /* For cache cleaning */
 
     uint64_t	 bluedot_ip_queue;
+
 #endif
